Fixed test code so that only one Spec exists per file
