+++ jruby-gradle-jar-plugin/src/test/groovy/com/github/jrubygradle/jar/JRubyJarPluginSpec.groovy
@@ -238,24 +238,3 @@ class JRubyJarPluginSpec extends Specification {
         prepareTask.dependencies.find { (it instanceof Configuration) && (it.name == jarTask.configuration) }
     }
 }
-
-
-class JRubyPluginInstanceSpec extends Specification {
-    JRubyJarPlugin plugin
-
-    def setup() {
-        plugin = new JRubyJarPlugin()
-    }
-
-    def "isJRubyVersionDeprecated()"() {
-        expect:
-        plugin.isJRubyVersionDeprecated(version) == expected
-
-        where:
-        version   | expected
-        '9.0.0.0' | false
-        '1.7.20'  | false
-        '1.7.11'  | true
-        '1.7.19'  | true
-    }
-}
